#2: refactor
